+++ tests/integration/query_execute_test.go
@@ -674,8 +674,8 @@ func TestQueryWideIntervalTypes(t *testing.T) {
 				CAST("PT20M34.56789S" AS Interval64),
 				CAST("PT20M34.56789S" AS Interval64),
 			;`,
-			expYdbValue: value.OptionalValue(value.Interval64ValueFromDuration(time.Duration(1234567890))),
-			expGoValue:  time.Duration(1234567890),
+			expYdbValue: value.OptionalValue(value.Interval64ValueFromDuration(1234567890 * time.Microsecond)),
+			expGoValue:  1234567890 * time.Microsecond,
 		},
 	} {
 		t.Run(tt.name, func(t *testing.T) {
@@ -997,5 +997,52 @@ func TestIssue1872QueryWarning(t *testing.T) {
 		require.Equal(t,
 			"Failed to convert type: Struct<'Amount':Decimal(22,9),'Id':Int32> to Struct<'Amount':Decimal(22,9)?,'Id':Uint64?>",
 			issueList[0].Issues[0].Issues[0].Message)
+	})
+}
+
+// https://github.com/ydb-platform/ydb-go-sdk/issues/1878
+func TestIssue1878ConcurrentResultSet(t *testing.T) {
+	ctx, cancel := context.WithCancel(xtest.Context(t))
+	defer cancel()
+	db, err := ydb.Open(ctx,
+		os.Getenv("YDB_CONNECTION_STRING"),
+		ydb.WithAccessTokenCredentials(os.Getenv("YDB_ACCESS_TOKEN_CREDENTIALS")),
+		ydb.WithTraceQuery(
+			log.Query(
+				log.Default(os.Stdout,
+					log.WithLogQuery(),
+					log.WithColoring(),
+					log.WithMinLevel(log.INFO),
+				),
+				trace.QueryEvents,
+			),
+		),
+	)
+	require.NoError(t, err)
+	t.Run("Select with enabled option", func(t *testing.T) {
+		q := db.Query()
+		res, err := q.Query(ctx, `
+				SELECT 1;
+				SELECT 2;
+				SELECT 3;
+				SELECT 4;
+				SELECT 5;
+	        `,
+			query.WithSyntax(query.SyntaxYQL),
+			query.WithIdempotent(),
+			query.WithConcurrentResultSets(true),
+		)
+		require.NoError(t, err)
+		rsCount := 0
+		for rs, err := range res.ResultSets(ctx) {
+			rsCount++
+			require.NoError(t, err)
+			row, err := rs.NextRow(ctx)
+			require.NoError(t, err)
+			require.Equal(t, 1, len(row.Values()))
+			require.EqualValues(t, rsCount, row.Values()[0])
+		}
+		require.NoError(t, res.Close(ctx))
+		require.Equal(t, 5, rsCount)
 	})
 }